+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-01T06:38:13.624Z",
+    "2026-05-01T09:39:49.168Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {