+++ packages/orq-rc/RELEASES.md
@@ -8779,3 +8779,13 @@ Based on:
 - [typescript v4.9.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.15 - packages/orq-rc
+
+## 2026-05-01 09:43:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.16 - packages/orq-rc